requirements.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.6
django-bootstrap3==10.0.1
django-ckeditor==5.4.0
django-js-asset==1.1.0
gunicorn==19.8.1
Markdown==2.6.11
Pillow==5.1.0
pkg-resources==0.0.0
pytz==2018.4
static3==0.7.0
